Implement order functionality with new services, repositories, and API endpoints.
- Added `OrderEntity`, `OrderProductEntity`, and `OrderResultEntity` along with corresponding repositories to support order management (`OrderRepository`, `OrderProductRepository`, `OrderResultRepository`).
- Introduced `OrderService` for creating and managing orders, invoice integration via Xendit API (`InvoiceService`), and webhook handling for payment updates.
- Created `OrderController` with endpoints for order retrieval, creation, execution, and webhook handling (`/api/v1/orders`).
- Updated DTOs and templates to support orders and invoices (`InvoiceRequest`, `InvoiceResponse`, `InvoiceWebhookRequest`, `OrderExecuteRequest`).
- Adjusted `CustomUserDetails` schema, encrypting password validation and user ID retrieval via security context.
- Modified `application.properties` to include Xendit API credentials for invoice management (`xendit.apikey.secret`, `xendit.webhook.secret`).
- Refactored `JwtUtils`, `CartRepository`, and `CustomUserDetailService` to integrate order and invoice services.

+++ src/main/kotlin/org/riezki/ecommerce/data/template/InvoiceRequest.kt
@@ -7,7 +7,7 @@ import com.fasterxml.jackson.databind.annotation.JsonNaming
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy::class)
 data class InvoiceRequest(
     val externalId: String = "external_id-1",
-    val amount: Int = 20000,
+    val amount: Long = 20000,
     val description: String = "Invoice demo #123",
     val invoiceDuration: Long = 86400L,
     val customer: Customer = Customer(),
